+++ src/main/java/SMU/BAMBOO/Hompage/domain/tag/service/TagService.java
@@ -1,7 +1,7 @@
-package SMU.BAMBOO.Hompage.domain.tag.entity.service;
+package SMU.BAMBOO.Hompage.domain.tag.service;
 
-import SMU.BAMBOO.Hompage.domain.tag.entity.dto.TagRequestDTO;
-import SMU.BAMBOO.Hompage.domain.tag.entity.dto.TagResponseDTO;
+import SMU.BAMBOO.Hompage.domain.tag.dto.TagRequestDTO;
+import SMU.BAMBOO.Hompage.domain.tag.dto.TagResponseDTO;
 
 import java.util.List;
